+++ src/m_misc.h
@@ -14,7 +14,7 @@ enum M_DataTypes { t_bool, t_char, t_uchar, t_short, t_ushort, t_long, t_ulong, 
 
 M_DataTypes M_DetermineType(string data);
  
-string M_ftoa(double number); // M_ntoa functions return a string class object 
+string M_ftoa(float number); // M_ntoa functions return a string class object 
 string M_itoa(long number); // with a character representation of its value
 string M_btoa(bool number);
 string M_itoah(unsigned short number, unsigned short placeCount, bool prefix = true); // hexadecimal output placecount = 4 for a word, 2 for a char
